UI tests refactoring
